requirements.txt
torch 
transformers 
pillow 
faiss-cpu 
gradio